requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
python-dotenv
pydantic
faiss-cpu
beautifulsoup4
